requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.26.1=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
apache-beam=2.53.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=2.2.1=pypi_0
comm=0.1.2=py38h06a4308_0
contourpy=1.1.1=pypi_0
crcmod=1.7=pypi_0
cycler=0.12.1=pypi_0
datasets=2.16.1=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
deprecated=1.2.14=pypi_0
dill=0.3.7=pypi_0
dm-tree=0.1.8=pypi_0
dnspython=2.5.0=pypi_0
docopt=0.6.2=pypi_0
executing=0.8.3=pyhd3eb1b0_0
fastavro=1.9.3=pypi_0
fasteners=0.19=pypi_0
filelock=3.13.1=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.47.2=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
gast=0.4.0=pypi_0
google-api-core=2.16.2=pypi_0
google-api-python-client=1.12.11=pypi_0
google-apitools=0.5.31=pypi_0
google-auth=2.27.0=pypi_0
google-auth-httplib2=0.1.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-cloud-aiplatform=1.40.0=pypi_0
google-cloud-bigquery=3.17.1=pypi_0
google-cloud-bigquery-storage=2.24.0=pypi_0
google-cloud-bigtable=2.22.0=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-datastore=2.19.0=pypi_0
google-cloud-dlp=3.15.0=pypi_0
google-cloud-language=2.13.0=pypi_0
google-cloud-pubsub=2.19.1=pypi_0
google-cloud-pubsublite=1.9.0=pypi_0
google-cloud-recommendations-ai=0.10.7=pypi_0
google-cloud-resource-manager=1.12.0=pypi_0
google-cloud-spanner=3.42.0=pypi_0
google-cloud-storage=2.14.0=pypi_0
google-cloud-videointelligence=2.13.0=pypi_0
google-cloud-vision=3.6.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-pasta=0.2.0=pypi_0
google-resumable-media=2.7.0=pypi_0
googleapis-common-protos=1.62.0=pypi_0
grpc-google-iam-v1=0.13.0=pypi_0
grpc-interceptor=0.15.4=pypi_0
grpcio=1.60.1=pypi_0
grpcio-status=1.48.2=pypi_0
h5py=3.10.0=pypi_0
hdfs=2.7.3=pypi_0
httplib2=0.22.0=pypi_0
huggingface-hub=0.20.3=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.0.1=py38h06a4308_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=7.0.1=hd3eb1b0_0
ipykernel=6.28.0=py38h06a4308_0
ipython=8.12.2=py38h06a4308_0
jedi=0.18.1=py38h06a4308_1
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
js2py=0.74=pypi_0
jsonpickle=3.0.2=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.0=py38h06a4308_0
jupyter_core=5.5.0=py38h06a4308_0
keras=2.13.1=pypi_0
keras-core=0.1.5=pypi_0
keras-nlp=0.6.1=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.4=pypi_0
matplotlib-inline=0.1.6=py38h06a4308_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.15=pypi_0
namex=0.0.7=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py38h06a4308_0
networkx=3.0=pypi_0
numpy=1.24.3=pypi_0
oauth2client=4.1.3=pypi_0
oauthlib=3.2.2=pypi_0
objsize=0.6.1=pypi_0
onnx=1.15.0=pypi_0
onnxconverter-common=1.13.0=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=h7f8727e_0
opt-einsum=3.3.0=pypi_0
orjson=3.9.12=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pandas=1.5.3=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.10.0=py38h06a4308_0
prompt-toolkit=3.0.43=py38h06a4308_0
proto-plus=1.23.0=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=10.0.1=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydot=1.4.2=pypi_0
pyfarmhash=0.3.2=pypi_0
pygments=2.15.1=py38h06a4308_1
pyjsparser=2.7.1=pypi_0
pymongo=4.6.1=pypi_0
pyparsing=3.1.1=pypi_0
pytesseract=0.3.10=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py38h6a678d5_0
readline=8.2=h5eee18b_0
referencing=0.33.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.17.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
sentencepiece=0.1.99=pypi_0
seqeval=1.2.2=pypi_0
setuptools=68.2.2=py38h06a4308_0
shapely=2.0.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=pypi_0
src=0.1.0=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.12=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.13.1=pypi_0
tensorflow-cpu=2.13.1=pypi_0
tensorflow-data-validation=1.14.0=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-hub=0.16.1=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
tensorflow-metadata=1.14.0=pypi_0
tensorflow-serving-api=2.13.1=pypi_0
tensorflow-text=2.13.0=pypi_0
termcolor=2.4.0=pypi_0
tf-keras=2.15.0=pypi_0
tf2onnx=1.16.1=pypi_0
tfx-bsl=1.14.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.1=pypi_0
torch=2.2.0+cpu=pypi_0
torchaudio=2.2.0+cpu=pypi_0
torchvision=0.17.0+cpu=pypi_0
tornado=6.3.3=py38h5eee18b_0
tqdm=4.66.1=pypi_0
traitlets=5.7.1=py38h06a4308_0
transformers=4.37.2=pypi_0
typing-extensions=4.8.0=pypi_0
typing_extensions=4.9.0=py38h06a4308_1
tzlocal=5.2=pypi_0
uritemplate=3.0.1=pypi_0
urllib3=2.2.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py38h06a4308_0
wrapt=1.16.0=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.5=h5eee18b_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstandard=0.22.0=pypi_0
 